requirements.txt
Pillow
matplotlib
numpy
opencv-python
scikit-learn
seaborn
streamlit
torch
transformers